(NPUP-15) Port negate operator to the dispatcher.
Porting the negate operator's implementation to the dispatcher.
